refactor: split parser.go into logical modules (Issue #78)
Split the monolithic parser.go (2523 lines) into focused modules:

New files created:
- expressions.go (583 lines): Expression parsing logic
  - parseExpression, parseAndExpression, parseComparisonExpression
  - parsePrimaryExpression, parseCaseExpression, parseSubquery

- window.go (285 lines): Window function support
  - parseFunctionCall, parseWindowSpec
  - parseWindowFrame, parseFrameBound, parseNullsClause

- grouping.go (156 lines): SQL-99 grouping operations
  - parseGroupingExpressionList, parseRollup
  - parseCube, parseGroupingSets

- select.go (524 lines): SELECT statement parsing
  - parseSelectStatement, parseSelectWithSetOperations
  - parseColumnDef, parseTableConstraint

- dml.go (512 lines): DML operations
  - parseInsertStatement, parseUpdateStatement
  - parseDeleteStatement, parseMergeStatement
  - parseMergeWhenClause, parseMergeAction

- cte.go (175 lines): Common Table Expressions
  - parseWithStatement, parseCommonTableExpr
  - parseMainStatementAfterWith

Reduced parser.go to 356 lines (core):
- Parser struct and Parse/ParseContext methods
- Basic utilities: advance, peekToken, matchToken
- parseStatement switch
- Helper functions for identifiers and table references

Benefits:
- Each file < 600 lines (was 2523)
- Logical grouping of related functions
- Easier navigation and maintenance
- No breaking changes - all tests pass
