+++ pkg/sql/parser/parser.go
@@ -295,2011 +295,6 @@ func (p *Parser) parseStringLiteral() string {
 	return value
 }
 
-// parseExpression parses an expression with OR operators (lowest precedence)
-func (p *Parser) parseExpression() (ast.Expression, error) {
-	// Check context if available
-	if p.ctx != nil {
-		if err := p.ctx.Err(); err != nil {
-			return nil, fmt.Errorf("parsing cancelled: %w", err)
-		}
-	}
-
-	// Check recursion depth to prevent stack overflow
-	p.depth++
-	defer func() { p.depth-- }()
-
-	if p.depth > MaxRecursionDepth {
-		return nil, fmt.Errorf("maximum recursion depth exceeded (%d) - expression too deeply nested", MaxRecursionDepth)
-	}
-
-	// Start by parsing AND expressions (higher precedence)
-	left, err := p.parseAndExpression()
-	if err != nil {
-		return nil, err
-	}
-
-	// Handle OR operators (lowest precedence, left-associative)
-	for p.currentToken.Type == "OR" {
-		operator := p.currentToken.Literal
-		p.advance() // Consume OR
-
-		right, err := p.parseAndExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		left = &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    right,
-		}
-	}
-
-	return left, nil
-}
-
-// parseAndExpression parses an expression with AND operators (middle precedence)
-func (p *Parser) parseAndExpression() (ast.Expression, error) {
-	// Parse comparison expressions (higher precedence)
-	left, err := p.parseComparisonExpression()
-	if err != nil {
-		return nil, err
-	}
-
-	// Handle AND operators (middle precedence, left-associative)
-	for p.currentToken.Type == "AND" {
-		operator := p.currentToken.Literal
-		p.advance() // Consume AND
-
-		right, err := p.parseComparisonExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		left = &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    right,
-		}
-	}
-
-	return left, nil
-}
-
-// parseComparisonExpression parses an expression with comparison operators (highest precedence)
-func (p *Parser) parseComparisonExpression() (ast.Expression, error) {
-	// Parse the left side (primary expression)
-	left, err := p.parsePrimaryExpression()
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for NOT prefix for BETWEEN, LIKE, IN operators
-	// Only consume NOT if followed by BETWEEN, LIKE, ILIKE, or IN
-	// This prevents breaking cases like: WHERE NOT active AND name LIKE '%'
-	notPrefix := false
-	if p.currentToken.Type == "NOT" {
-		nextToken := p.peekToken()
-		if nextToken.Type == "BETWEEN" || nextToken.Type == "LIKE" || nextToken.Type == "ILIKE" || nextToken.Type == "IN" {
-			notPrefix = true
-			p.advance() // Consume NOT only if followed by valid operator
-		}
-	}
-
-	// Check for BETWEEN operator
-	if p.currentToken.Type == "BETWEEN" {
-		p.advance() // Consume BETWEEN
-
-		// Parse lower bound
-		lower, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse BETWEEN lower bound: %w", err)
-		}
-
-		// Expect AND keyword
-		if p.currentToken.Type != "AND" {
-			return nil, p.expectedError("AND")
-		}
-		p.advance() // Consume AND
-
-		// Parse upper bound
-		upper, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse BETWEEN upper bound: %w", err)
-		}
-
-		return &ast.BetweenExpression{
-			Expr:  left,
-			Lower: lower,
-			Upper: upper,
-			Not:   notPrefix,
-		}, nil
-	}
-
-	// Check for LIKE/ILIKE operator
-	if p.currentToken.Type == "LIKE" || p.currentToken.Type == "ILIKE" {
-		operator := p.currentToken.Literal
-		p.advance() // Consume LIKE/ILIKE
-
-		// Parse pattern
-		pattern, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse LIKE pattern: %w", err)
-		}
-
-		return &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    pattern,
-			Not:      notPrefix,
-		}, nil
-	}
-
-	// Check for IN operator
-	if p.currentToken.Type == "IN" {
-		p.advance() // Consume IN
-
-		// Expect opening parenthesis
-		if p.currentToken.Type != "(" {
-			return nil, p.expectedError("(")
-		}
-		p.advance() // Consume (
-
-		// Check if this is a subquery (starts with SELECT or WITH)
-		if p.currentToken.Type == "SELECT" || p.currentToken.Type == "WITH" {
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse IN subquery: %w", err)
-			}
-
-			// Expect closing parenthesis
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			return &ast.InExpression{
-				Expr:     left,
-				Subquery: subquery,
-				Not:      notPrefix,
-			}, nil
-		}
-
-		// Parse value list
-		var values []ast.Expression
-		for {
-			value, err := p.parseExpression()
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse IN value: %w", err)
-			}
-			values = append(values, value)
-
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-			} else if p.currentToken.Type == ")" {
-				break
-			} else {
-				return nil, p.expectedError(", or )")
-			}
-		}
-		p.advance() // Consume )
-
-		return &ast.InExpression{
-			Expr: left,
-			List: values,
-			Not:  notPrefix,
-		}, nil
-	}
-
-	// If NOT was consumed but no BETWEEN/LIKE/IN follows, we need to handle this case
-	// Put back the NOT by creating a NOT expression with left as the operand
-	if notPrefix {
-		return nil, fmt.Errorf("expected BETWEEN, LIKE, or IN after NOT")
-	}
-
-	// Check for IS NULL / IS NOT NULL
-	if p.currentToken.Type == "IS" {
-		p.advance() // Consume IS
-
-		isNot := false
-		if p.currentToken.Type == "NOT" {
-			isNot = true
-			p.advance() // Consume NOT
-		}
-
-		if p.currentToken.Type == "NULL" {
-			p.advance() // Consume NULL
-			return &ast.BinaryExpression{
-				Left:     left,
-				Operator: "IS NULL",
-				Right:    &ast.LiteralValue{Value: nil, Type: "null"},
-				Not:      isNot,
-			}, nil
-		}
-
-		return nil, p.expectedError("NULL")
-	}
-
-	// Check if this is a comparison binary expression
-	if p.currentToken.Type == "=" || p.currentToken.Type == "<" ||
-		p.currentToken.Type == ">" || p.currentToken.Type == "!=" ||
-		p.currentToken.Type == "<=" || p.currentToken.Type == ">=" ||
-		p.currentToken.Type == "<>" {
-		// Save the operator
-		operator := p.currentToken.Literal
-		p.advance()
-
-		// Check for ANY/ALL subquery operators
-		if p.currentToken.Type == "ANY" || p.currentToken.Type == "ALL" {
-			quantifier := p.currentToken.Type
-			p.advance() // Consume ANY/ALL
-
-			// Expect opening parenthesis
-			if p.currentToken.Type != "(" {
-				return nil, p.expectedError("(")
-			}
-			p.advance() // Consume (
-
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse %s subquery: %w", quantifier, err)
-			}
-
-			// Expect closing parenthesis
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			if quantifier == "ANY" {
-				return &ast.AnyExpression{
-					Expr:     left,
-					Operator: operator,
-					Subquery: subquery,
-				}, nil
-			}
-			return &ast.AllExpression{
-				Expr:     left,
-				Operator: operator,
-				Subquery: subquery,
-			}, nil
-		}
-
-		// Parse the right side of the expression
-		right, err := p.parsePrimaryExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		// Create a binary expression
-		return &ast.BinaryExpression{
-			Left:     left,
-			Operator: operator,
-			Right:    right,
-		}, nil
-	}
-
-	return left, nil
-}
-
-// parsePrimaryExpression parses a primary expression (literals, identifiers, function calls)
-func (p *Parser) parsePrimaryExpression() (ast.Expression, error) {
-	switch p.currentToken.Type {
-	case "CASE":
-		// Handle CASE expressions (both simple and searched forms)
-		return p.parseCaseExpression()
-
-	case "IDENT":
-		// Handle identifiers and function calls
-		identName := p.currentToken.Literal
-		p.advance()
-
-		// Check for function call (identifier followed by parentheses)
-		if p.currentToken.Type == "(" {
-			// This is a function call
-			funcCall, err := p.parseFunctionCall(identName)
-			if err != nil {
-				return nil, err
-			}
-			return funcCall, nil
-		}
-
-		// Handle regular identifier or qualified identifier (table.column)
-		ident := &ast.Identifier{Name: identName}
-
-		// Check for qualified identifier (table.column)
-		if p.currentToken.Type == "." {
-			p.advance() // Consume .
-			if p.currentToken.Type != "IDENT" {
-				return nil, p.expectedError("identifier after .")
-			}
-			// Create a qualified identifier
-			ident = &ast.Identifier{
-				Table: ident.Name,
-				Name:  p.currentToken.Literal,
-			}
-			p.advance()
-		}
-
-		return ident, nil
-
-	case "*":
-		// Handle asterisk (e.g., in COUNT(*) or SELECT *)
-		p.advance()
-		return &ast.Identifier{Name: "*"}, nil
-
-	case "STRING":
-		// Handle string literals
-		value := p.currentToken.Literal
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "string"}, nil
-
-	case "INT":
-		// Handle integer literals
-		value := p.currentToken.Literal
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "int"}, nil
-
-	case "FLOAT":
-		// Handle float literals
-		value := p.currentToken.Literal
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "float"}, nil
-
-	case "TRUE", "FALSE":
-		// Handle boolean literals
-		value := p.currentToken.Literal
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "bool"}, nil
-
-	case "PLACEHOLDER":
-		// Handle SQL placeholders (e.g., $1, $2 for PostgreSQL; @param for SQL Server)
-		value := p.currentToken.Literal
-		p.advance()
-		return &ast.LiteralValue{Value: value, Type: "placeholder"}, nil
-
-	case "NULL":
-		// Handle NULL literal
-		p.advance()
-		return &ast.LiteralValue{Value: nil, Type: "null"}, nil
-
-	case "(":
-		// Handle parenthesized expression or subquery
-		p.advance() // Consume (
-
-		// Check if this is a subquery (starts with SELECT or WITH)
-		if p.currentToken.Type == "SELECT" || p.currentToken.Type == "WITH" {
-			// Parse subquery
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse subquery: %w", err)
-			}
-			// Expect closing parenthesis
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-			return &ast.SubqueryExpression{Subquery: subquery}, nil
-		}
-
-		// Regular parenthesized expression
-		expr, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		// Expect closing parenthesis
-		if p.currentToken.Type != ")" {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-		return expr, nil
-
-	case "EXISTS":
-		// Handle EXISTS (subquery)
-		p.advance() // Consume EXISTS
-
-		// Expect opening parenthesis
-		if p.currentToken.Type != "(" {
-			return nil, p.expectedError("(")
-		}
-		p.advance() // Consume (
-
-		// Parse the subquery
-		subquery, err := p.parseSubquery()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse EXISTS subquery: %w", err)
-		}
-
-		// Expect closing parenthesis
-		if p.currentToken.Type != ")" {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-
-		return &ast.ExistsExpression{Subquery: subquery}, nil
-
-	case "NOT":
-		// Handle NOT expression (NOT EXISTS, NOT boolean)
-		p.advance() // Consume NOT
-
-		if p.currentToken.Type == "EXISTS" {
-			// NOT EXISTS (subquery)
-			p.advance() // Consume EXISTS
-
-			if p.currentToken.Type != "(" {
-				return nil, p.expectedError("(")
-			}
-			p.advance() // Consume (
-
-			subquery, err := p.parseSubquery()
-			if err != nil {
-				return nil, fmt.Errorf("failed to parse NOT EXISTS subquery: %w", err)
-			}
-
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-
-			// Return NOT EXISTS as a BinaryExpression with NOT flag
-			return &ast.BinaryExpression{
-				Left:     &ast.ExistsExpression{Subquery: subquery},
-				Operator: "NOT",
-				Right:    nil,
-				Not:      true,
-			}, nil
-		}
-
-		// NOT followed by other expression (boolean negation)
-		// Parse at comparison level for proper precedence: NOT (a > b), NOT active
-		expr, err := p.parseComparisonExpression()
-		if err != nil {
-			return nil, err
-		}
-		return &ast.UnaryExpression{
-			Operator: ast.Not,
-			Expr:     expr,
-		}, nil
-
-	default:
-		return nil, fmt.Errorf("unexpected token: %s", p.currentToken.Type)
-	}
-}
-
-// parseCaseExpression parses a CASE expression (both simple and searched forms)
-//
-// Simple CASE: CASE expr WHEN value THEN result ... [ELSE result] END
-// Searched CASE: CASE WHEN condition THEN result ... [ELSE result] END
-func (p *Parser) parseCaseExpression() (*ast.CaseExpression, error) {
-	p.advance() // Consume CASE
-
-	caseExpr := &ast.CaseExpression{
-		WhenClauses: make([]ast.WhenClause, 0),
-	}
-
-	// Check if this is a simple CASE (has a value expression) or searched CASE (no value)
-	// Simple CASE: CASE expr WHEN value THEN result
-	// Searched CASE: CASE WHEN condition THEN result
-	if p.currentToken.Type != "WHEN" {
-		// This is a simple CASE - parse the value expression
-		value, err := p.parseExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse CASE value: %w", err)
-		}
-		caseExpr.Value = value
-	}
-
-	// Parse WHEN clauses (at least one required)
-	for p.currentToken.Type == "WHEN" {
-		p.advance() // Consume WHEN
-
-		// Parse the condition/value expression
-		condition, err := p.parseExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse WHEN condition: %w", err)
-		}
-
-		// Expect THEN keyword
-		if p.currentToken.Type != "THEN" {
-			return nil, p.expectedError("THEN")
-		}
-		p.advance() // Consume THEN
-
-		// Parse the result expression
-		result, err := p.parseExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse THEN result: %w", err)
-		}
-
-		caseExpr.WhenClauses = append(caseExpr.WhenClauses, ast.WhenClause{
-			Condition: condition,
-			Result:    result,
-		})
-	}
-
-	// Check that we have at least one WHEN clause
-	if len(caseExpr.WhenClauses) == 0 {
-		return nil, fmt.Errorf("CASE expression requires at least one WHEN clause")
-	}
-
-	// Parse optional ELSE clause
-	if p.currentToken.Type == "ELSE" {
-		p.advance() // Consume ELSE
-
-		elseResult, err := p.parseExpression()
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse ELSE result: %w", err)
-		}
-		caseExpr.ElseClause = elseResult
-	}
-
-	// Expect END keyword
-	if p.currentToken.Type != "END" {
-		return nil, p.expectedError("END")
-	}
-	p.advance() // Consume END
-
-	return caseExpr, nil
-}
-
-// parseSubquery parses a subquery (SELECT or WITH statement).
-// Expects current token to be SELECT or WITH.
-func (p *Parser) parseSubquery() (ast.Statement, error) {
-	if p.currentToken.Type == "WITH" {
-		// WITH statement handles its own token consumption
-		return p.parseWithStatement()
-	}
-
-	if p.currentToken.Type == "SELECT" {
-		p.advance() // Consume SELECT
-		return p.parseSelectWithSetOperations()
-	}
-
-	return nil, fmt.Errorf("expected SELECT or WITH, got %s", p.currentToken.Type)
-}
-
-// parseFunctionCall parses a function call with optional OVER clause for window functions.
-//
-// Examples:
-//
-//	COUNT(*) -> regular aggregate function
-//	ROW_NUMBER() OVER (ORDER BY id) -> window function with OVER clause
-//	SUM(salary) OVER (PARTITION BY dept ORDER BY date ROWS UNBOUNDED PRECEDING) -> window function with frame
-func (p *Parser) parseFunctionCall(funcName string) (*ast.FunctionCall, error) {
-	// Expect opening parenthesis
-	if p.currentToken.Type != "(" {
-		return nil, p.expectedError("(")
-	}
-	p.advance() // Consume (
-
-	// Parse function arguments
-	var arguments []ast.Expression
-	var distinct bool
-
-	// Check for DISTINCT keyword
-	if p.currentToken.Type == "DISTINCT" {
-		distinct = true
-		p.advance()
-	}
-
-	// Parse arguments if not empty
-	if p.currentToken.Type != ")" {
-		for {
-			arg, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-			arguments = append(arguments, arg)
-
-			// Check for comma or end of arguments
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-			} else if p.currentToken.Type == ")" {
-				break
-			} else {
-				return nil, p.expectedError(", or )")
-			}
-		}
-	}
-
-	// Expect closing parenthesis
-	if p.currentToken.Type != ")" {
-		return nil, p.expectedError(")")
-	}
-	p.advance() // Consume )
-
-	// Create function call
-	funcCall := &ast.FunctionCall{
-		Name:      funcName,
-		Arguments: arguments,
-		Distinct:  distinct,
-	}
-
-	// Check for OVER clause (window function)
-	if p.currentToken.Type == "OVER" {
-		p.advance() // Consume OVER
-
-		windowSpec, err := p.parseWindowSpec()
-		if err != nil {
-			return nil, err
-		}
-		funcCall.Over = windowSpec
-	}
-
-	return funcCall, nil
-}
-
-// parseWindowSpec parses a window specification (PARTITION BY, ORDER BY, frame clause)
-func (p *Parser) parseWindowSpec() (*ast.WindowSpec, error) {
-	// Expect opening parenthesis
-	if p.currentToken.Type != "(" {
-		return nil, p.expectedError("(")
-	}
-	p.advance() // Consume (
-
-	windowSpec := &ast.WindowSpec{}
-
-	// Parse PARTITION BY clause
-	if p.currentToken.Type == "PARTITION" {
-		p.advance() // Consume PARTITION
-		if p.currentToken.Type != "BY" {
-			return nil, p.expectedError("BY after PARTITION")
-		}
-		p.advance() // Consume BY
-
-		// Parse partition expressions
-		for {
-			expr, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-			windowSpec.PartitionBy = append(windowSpec.PartitionBy, expr)
-
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-			} else {
-				break
-			}
-		}
-	}
-
-	// Parse ORDER BY clause
-	if p.currentToken.Type == "ORDER" {
-		p.advance() // Consume ORDER
-		if p.currentToken.Type != "BY" {
-			return nil, p.expectedError("BY after ORDER")
-		}
-		p.advance() // Consume BY
-
-		// Parse order expressions
-		for {
-			expr, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-
-			// Create OrderByExpression with defaults
-			orderByExpr := ast.OrderByExpression{
-				Expression: expr,
-				Ascending:  true, // Default to ASC
-				NullsFirst: nil,  // Default behavior (database-specific)
-			}
-
-			// Check for ASC/DESC after the expression
-			if p.currentToken.Type == "ASC" {
-				orderByExpr.Ascending = true
-				p.advance() // Consume ASC
-			} else if p.currentToken.Type == "DESC" {
-				orderByExpr.Ascending = false
-				p.advance() // Consume DESC
-			}
-
-			// Check for NULLS FIRST/LAST
-			nullsFirst, err := p.parseNullsClause()
-			if err != nil {
-				return nil, err
-			}
-			orderByExpr.NullsFirst = nullsFirst
-
-			windowSpec.OrderBy = append(windowSpec.OrderBy, orderByExpr)
-
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-			} else {
-				break
-			}
-		}
-	}
-
-	// Parse frame clause (ROWS/RANGE with bounds)
-	if p.currentToken.Type == "ROWS" || p.currentToken.Type == "RANGE" {
-		frameType := p.currentToken.Literal
-		p.advance() // Consume ROWS/RANGE
-
-		frameClause, err := p.parseWindowFrame(frameType)
-		if err != nil {
-			return nil, err
-		}
-		windowSpec.FrameClause = frameClause
-	}
-
-	// Expect closing parenthesis
-	if p.currentToken.Type != ")" {
-		return nil, p.expectedError(")")
-	}
-	p.advance() // Consume )
-
-	return windowSpec, nil
-}
-
-// parseWindowFrame parses a window frame clause
-func (p *Parser) parseWindowFrame(frameType string) (*ast.WindowFrame, error) {
-	frame := &ast.WindowFrame{
-		Type: frameType,
-	}
-
-	// Parse frame bounds
-	if p.currentToken.Type == "BETWEEN" {
-		p.advance() // Consume BETWEEN
-
-		// Parse start bound
-		startBound, err := p.parseFrameBound()
-		if err != nil {
-			return nil, err
-		}
-		frame.Start = *startBound
-
-		// Expect AND
-		if p.currentToken.Type != "AND" {
-			return nil, p.expectedError("AND")
-		}
-		p.advance() // Consume AND
-
-		// Parse end bound
-		endBound, err := p.parseFrameBound()
-		if err != nil {
-			return nil, err
-		}
-		frame.End = endBound
-	} else {
-		// Single bound (implies CURRENT ROW as end)
-		startBound, err := p.parseFrameBound()
-		if err != nil {
-			return nil, err
-		}
-		frame.Start = *startBound
-		// End is nil for single bound
-	}
-
-	return frame, nil
-}
-
-// parseFrameBound parses a window frame bound
-func (p *Parser) parseFrameBound() (*ast.WindowFrameBound, error) {
-	bound := &ast.WindowFrameBound{}
-
-	if p.currentToken.Type == "UNBOUNDED" {
-		p.advance() // Consume UNBOUNDED
-		if p.currentToken.Type == "PRECEDING" {
-			bound.Type = "UNBOUNDED PRECEDING"
-			p.advance() // Consume PRECEDING
-		} else if p.currentToken.Type == "FOLLOWING" {
-			bound.Type = "UNBOUNDED FOLLOWING"
-			p.advance() // Consume FOLLOWING
-		} else {
-			return nil, p.expectedError("PRECEDING or FOLLOWING after UNBOUNDED")
-		}
-	} else if p.currentToken.Type == "CURRENT" {
-		p.advance() // Consume CURRENT
-		if p.currentToken.Type != "ROW" {
-			return nil, p.expectedError("ROW after CURRENT")
-		}
-		bound.Type = "CURRENT ROW"
-		p.advance() // Consume ROW
-	} else {
-		// Numeric bound
-		expr, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-		bound.Value = expr
-
-		if p.currentToken.Type == "PRECEDING" {
-			bound.Type = "PRECEDING"
-			p.advance() // Consume PRECEDING
-		} else if p.currentToken.Type == "FOLLOWING" {
-			bound.Type = "FOLLOWING"
-			p.advance() // Consume FOLLOWING
-		} else {
-			return nil, p.expectedError("PRECEDING or FOLLOWING after numeric value")
-		}
-	}
-
-	return bound, nil
-}
-
-// parseNullsClause parses the optional NULLS FIRST/LAST clause in ORDER BY expressions.
-// Returns a pointer to bool indicating null ordering: true for NULLS FIRST, false for NULLS LAST, nil if not specified.
-func (p *Parser) parseNullsClause() (*bool, error) {
-	if p.currentToken.Type == "NULLS" {
-		p.advance() // Consume NULLS
-		if p.currentToken.Type == "FIRST" {
-			t := true
-			p.advance() // Consume FIRST
-			return &t, nil
-		} else if p.currentToken.Type == "LAST" {
-			f := false
-			p.advance() // Consume LAST
-			return &f, nil
-		} else {
-			return nil, p.expectedError("FIRST or LAST after NULLS")
-		}
-	}
-	return nil, nil
-}
-
-// parseGroupingExpressionList parses a parenthesized, comma-separated list of expressions
-// used by ROLLUP and CUBE. Returns error if the list is empty.
-func (p *Parser) parseGroupingExpressionList(keyword string) ([]ast.Expression, error) {
-	if p.currentToken.Type != "(" {
-		return nil, p.expectedError("( after " + keyword)
-	}
-	p.advance() // Consume (
-
-	// Check for empty list - not allowed for ROLLUP/CUBE
-	if p.currentToken.Type == ")" {
-		return nil, fmt.Errorf("parsing failed: %s requires at least one expression", keyword)
-	}
-
-	// Parse comma-separated expressions
-	expressions := make([]ast.Expression, 0)
-	for {
-		expr, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-		expressions = append(expressions, expr)
-
-		// Check for comma (more expressions) or closing paren
-		if p.currentToken.Type == ")" {
-			break
-		}
-		if p.currentToken.Type != "," {
-			return nil, p.expectedError(", or ) in " + keyword)
-		}
-		p.advance() // Consume comma
-	}
-	p.advance() // Consume )
-
-	return expressions, nil
-}
-
-// parseRollup parses ROLLUP(col1, col2, ...) in GROUP BY clause
-// ROLLUP generates hierarchical grouping sets from right to left
-// Example: ROLLUP(a, b, c) generates: (a, b, c), (a, b), (a), ()
-func (p *Parser) parseRollup() (*ast.RollupExpression, error) {
-	p.advance() // Consume ROLLUP
-
-	expressions, err := p.parseGroupingExpressionList("ROLLUP")
-	if err != nil {
-		return nil, err
-	}
-
-	return &ast.RollupExpression{
-		Expressions: expressions,
-	}, nil
-}
-
-// parseCube parses CUBE(col1, col2, ...) in GROUP BY clause
-// CUBE generates all possible combinations of grouping sets
-// Example: CUBE(a, b) generates: (a, b), (a), (b), ()
-func (p *Parser) parseCube() (*ast.CubeExpression, error) {
-	p.advance() // Consume CUBE
-
-	expressions, err := p.parseGroupingExpressionList("CUBE")
-	if err != nil {
-		return nil, err
-	}
-
-	return &ast.CubeExpression{
-		Expressions: expressions,
-	}, nil
-}
-
-// parseGroupingSets parses GROUPING SETS(...) in GROUP BY clause
-// Allows explicit specification of grouping sets
-// Example: GROUPING SETS((a, b), (a), ()) generates exactly those three grouping sets
-func (p *Parser) parseGroupingSets() (*ast.GroupingSetsExpression, error) {
-	// Handle both "GROUPING SETS" as compound keyword or separate tokens
-	if p.currentToken.Literal == "GROUPING SETS" {
-		p.advance() // Consume "GROUPING SETS" compound token
-	} else if p.currentToken.Type == "GROUPING" {
-		p.advance() // Consume GROUPING
-		if p.currentToken.Type != "SETS" {
-			return nil, p.expectedError("SETS after GROUPING")
-		}
-		p.advance() // Consume SETS
-	}
-
-	if p.currentToken.Type != "(" {
-		return nil, p.expectedError("( after GROUPING SETS")
-	}
-	p.advance() // Consume (
-
-	// Parse comma-separated grouping sets
-	sets := make([][]ast.Expression, 0)
-	for {
-		// Each set is either:
-		// 1. A parenthesized list: (col1, col2)
-		// 2. An empty set: ()
-		// 3. A single column without parens: col1 (treated as (col1))
-
-		var set []ast.Expression
-		if p.currentToken.Type == "(" {
-			p.advance() // Consume (
-			// Parse expressions in this set
-			set = make([]ast.Expression, 0)
-			// Handle empty set: ()
-			if p.currentToken.Type != ")" {
-				for {
-					expr, err := p.parseExpression()
-					if err != nil {
-						return nil, err
-					}
-					set = append(set, expr)
-
-					if p.currentToken.Type == ")" {
-						break
-					}
-					if p.currentToken.Type != "," {
-						return nil, p.expectedError(", or ) in grouping set")
-					}
-					p.advance() // Consume comma
-				}
-			}
-			p.advance() // Consume )
-		} else {
-			// Single column without parens
-			expr, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-			set = []ast.Expression{expr}
-		}
-		sets = append(sets, set)
-
-		// Check for comma (more sets) or closing paren
-		if p.currentToken.Type == ")" {
-			break
-		}
-		if p.currentToken.Type != "," {
-			return nil, p.expectedError(", or ) in GROUPING SETS")
-		}
-		p.advance() // Consume comma
-	}
-	p.advance() // Consume )
-
-	return &ast.GroupingSetsExpression{
-		Sets: sets,
-	}, nil
-}
-
-// parseColumnDef parses a column definition
-func (p *Parser) parseColumnDef() (*ast.ColumnDef, error) {
-	name := p.parseIdent()
-	if name == nil {
-		return nil, fmt.Errorf("expected column name")
-	}
-
-	dataType := p.parseIdent()
-	if dataType == nil {
-		return nil, fmt.Errorf("expected data type")
-	}
-
-	colDef := &ast.ColumnDef{
-		Name: name.Name,
-		Type: dataType.Name,
-	}
-
-	return colDef, nil
-}
-
-// parseTableConstraint parses a table constraint
-func (p *Parser) parseTableConstraint() (*ast.TableConstraint, error) {
-	name := p.parseIdent()
-	if name == nil {
-		return nil, fmt.Errorf("expected constraint name")
-	}
-
-	constraint := &ast.TableConstraint{
-		Name: name.Name,
-	}
-
-	return constraint, nil
-}
-
-// parseSelectStatement parses a SELECT statement
-func (p *Parser) parseSelectStatement() (ast.Statement, error) {
-	// We've already consumed the SELECT token in matchToken
-
-	// Parse columns
-	columns := make([]ast.Expression, 0)
-	for {
-		// Handle * as a special case
-		if p.currentToken.Type == "*" {
-			columns = append(columns, &ast.Identifier{Name: "*"})
-			p.advance()
-		} else {
-			// Use parseExpression to handle all types including function calls
-			expr, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-
-			// Check for optional column alias (AS alias_name)
-			if p.currentToken.Type == "AS" {
-				p.advance() // Consume AS
-				if p.currentToken.Type != "IDENT" {
-					return nil, p.expectedError("alias name after AS")
-				}
-				// Consume the alias name (for now we don't store it in AST)
-				p.advance()
-			}
-
-			columns = append(columns, expr)
-		}
-
-		// Check if there are more columns
-		if p.currentToken.Type != "," {
-			break
-		}
-		p.advance() // Consume comma
-	}
-
-	// Parse FROM clause (optional to support SELECT without FROM like "SELECT 1")
-	if p.currentToken.Type != "FROM" && p.currentToken.Type != "EOF" && p.currentToken.Type != ";" {
-		// If not FROM, EOF, or semicolon, it's likely an error
-		return nil, p.expectedError("FROM, semicolon, or end of statement")
-	}
-
-	var tableName string
-	var tables []ast.TableReference
-	var joins []ast.JoinClause
-
-	if p.currentToken.Type == "FROM" {
-		p.advance() // Consume FROM
-
-		// Parse table name
-		if p.currentToken.Type != "IDENT" {
-			return nil, p.expectedError("table name")
-		}
-		tableName = p.currentToken.Literal
-		p.advance()
-
-		// Create table reference
-		tableRef := ast.TableReference{
-			Name: tableName,
-		}
-
-		// Check for table alias
-		if p.currentToken.Type == "IDENT" || p.currentToken.Type == "AS" {
-			if p.currentToken.Type == "AS" {
-				p.advance() // Consume AS
-				if p.currentToken.Type != "IDENT" {
-					return nil, p.expectedError("alias after AS")
-				}
-			}
-			if p.currentToken.Type == "IDENT" {
-				tableRef.Alias = p.currentToken.Literal
-				p.advance()
-			}
-		}
-
-		// Create tables list for FROM clause
-		tables = []ast.TableReference{tableRef}
-
-		// Parse JOIN clauses if present
-		joins = []ast.JoinClause{}
-		for p.isJoinKeyword() {
-			// Determine JOIN type
-			joinType := "INNER" // Default
-
-			if p.currentToken.Type == "LEFT" {
-				joinType = "LEFT"
-				p.advance()
-				if p.currentToken.Type == "OUTER" {
-					p.advance() // Optional OUTER keyword
-				}
-			} else if p.currentToken.Type == "RIGHT" {
-				joinType = "RIGHT"
-				p.advance()
-				if p.currentToken.Type == "OUTER" {
-					p.advance() // Optional OUTER keyword
-				}
-			} else if p.currentToken.Type == "FULL" {
-				joinType = "FULL"
-				p.advance()
-				if p.currentToken.Type == "OUTER" {
-					p.advance() // Optional OUTER keyword
-				}
-			} else if p.currentToken.Type == "INNER" {
-				joinType = "INNER"
-				p.advance()
-			} else if p.currentToken.Type == "CROSS" {
-				joinType = "CROSS"
-				p.advance()
-			}
-
-			// Expect JOIN keyword
-			if p.currentToken.Type != "JOIN" {
-				return nil, fmt.Errorf("expected JOIN after %s, got %s", joinType, p.currentToken.Type)
-			}
-			p.advance() // Consume JOIN
-
-			// Parse joined table name
-			if p.currentToken.Type != "IDENT" {
-				return nil, fmt.Errorf("expected table name after %s JOIN, got %s", joinType, p.currentToken.Type)
-			}
-			joinedTableName := p.currentToken.Literal
-			p.advance()
-
-			// Create joined table reference
-			joinedTableRef := ast.TableReference{
-				Name: joinedTableName,
-			}
-
-			// Check for table alias
-			if p.currentToken.Type == "IDENT" || p.currentToken.Type == "AS" {
-				if p.currentToken.Type == "AS" {
-					p.advance() // Consume AS
-					if p.currentToken.Type != "IDENT" {
-						return nil, p.expectedError("alias after AS")
-					}
-				}
-				if p.currentToken.Type == "IDENT" {
-					joinedTableRef.Alias = p.currentToken.Literal
-					p.advance()
-				}
-			}
-
-			// Parse join condition (ON or USING)
-			var joinCondition ast.Expression
-
-			// CROSS JOIN doesn't require ON clause
-			if joinType != "CROSS" {
-				if p.currentToken.Type == "ON" {
-					p.advance() // Consume ON
-
-					// Parse join condition
-					cond, err := p.parseExpression()
-					if err != nil {
-						return nil, fmt.Errorf("error parsing ON condition for %s JOIN: %v", joinType, err)
-					}
-					joinCondition = cond
-				} else if p.currentToken.Type == "USING" {
-					p.advance() // Consume USING
-
-					// Parse column list in parentheses
-					if p.currentToken.Type != "(" {
-						return nil, p.expectedError("( after USING")
-					}
-					p.advance()
-
-					// Parse comma-separated column list for USING clause
-					// Supports both single column: USING (id)
-					// and multi-column: USING (id, name, category)
-					var usingColumns []ast.Expression
-
-					for {
-						// Parse column name
-						if p.currentToken.Type != "IDENT" {
-							return nil, p.expectedError("column name in USING")
-						}
-						usingColumns = append(usingColumns, &ast.Identifier{Name: p.currentToken.Literal})
-						p.advance()
-
-						// Check for comma (more columns)
-						if p.currentToken.Type == "," {
-							p.advance() // Consume comma
-							continue
-						}
-						break
-					}
-
-					// Check for closing parenthesis
-					if p.currentToken.Type != ")" {
-						return nil, p.expectedError(") after USING column list")
-					}
-					p.advance()
-
-					// Store as single identifier for single column (backward compatibility)
-					// or as ListExpression for multiple columns
-					if len(usingColumns) == 1 {
-						joinCondition = usingColumns[0]
-					} else {
-						joinCondition = &ast.ListExpression{Values: usingColumns}
-					}
-				} else if joinType != "NATURAL" {
-					return nil, p.expectedError("ON or USING")
-				}
-			}
-
-			// Create join clause with proper tree relationships
-			// For SQL: FROM A JOIN B JOIN C (equivalent to (A JOIN B) JOIN C)
-			var leftTable ast.TableReference
-			if len(joins) == 0 {
-				// First join: A JOIN B
-				leftTable = tableRef
-			} else {
-				// Subsequent joins: (previous result) JOIN C
-				// We represent this by using a synthetic table reference that indicates
-				// the left side is the result of previous joins
-				leftTable = ast.TableReference{
-					Name:  fmt.Sprintf("(%s_with_%d_joins)", tableRef.Name, len(joins)),
-					Alias: "",
-				}
-			}
-
-			joinClause := ast.JoinClause{
-				Type:      joinType,
-				Left:      leftTable,
-				Right:     joinedTableRef,
-				Condition: joinCondition,
-			}
-
-			// Add join clause to joins list
-			joins = append(joins, joinClause)
-
-			// Note: We don't update tableRef here as each JOIN in the list
-			// represents a join with the accumulated result set
-		}
-	} // End of FROM clause parsing
-
-	// Initialize SELECT statement
-	selectStmt := &ast.SelectStatement{
-		Columns:   columns,
-		From:      tables,
-		Joins:     joins,
-		TableName: tableName, // Add this for compatibility with tests
-	}
-
-	// Parse WHERE clause if present
-	if p.currentToken.Type == "WHERE" {
-		p.advance() // Consume WHERE
-
-		// Parse WHERE condition
-		whereClause, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-
-		// Add WHERE clause to SELECT statement
-		selectStmt.Where = whereClause
-	}
-
-	// Parse GROUP BY clause if present
-	if p.currentToken.Type == "GROUP" {
-		p.advance() // Consume GROUP
-		if p.currentToken.Type != "BY" {
-			return nil, p.expectedError("BY after GROUP")
-		}
-		p.advance() // Consume BY
-
-		// Parse GROUP BY expressions (comma-separated list)
-		// Supports: regular expressions, ROLLUP, CUBE, GROUPING SETS
-		groupByExprs := make([]ast.Expression, 0)
-		for {
-			var expr ast.Expression
-			var err error
-
-			// Check for grouping operations: ROLLUP, CUBE, GROUPING SETS
-			// Note: GROUPING SETS may come as a compound keyword or separate tokens
-			if p.currentToken.Type == "ROLLUP" {
-				expr, err = p.parseRollup()
-			} else if p.currentToken.Type == "CUBE" {
-				expr, err = p.parseCube()
-			} else if p.currentToken.Literal == "GROUPING SETS" ||
-				(p.currentToken.Type == "GROUPING" && p.peekToken().Type == "SETS") {
-				expr, err = p.parseGroupingSets()
-			} else {
-				expr, err = p.parseExpression()
-			}
-
-			if err != nil {
-				return nil, err
-			}
-			groupByExprs = append(groupByExprs, expr)
-
-			// Check for comma (more expressions)
-			if p.currentToken.Type != "," {
-				break
-			}
-			p.advance() // Consume comma
-		}
-
-		// MySQL syntax support: GROUP BY col1, col2 WITH ROLLUP / WITH CUBE
-		// This is different from SQL-99 GROUP BY ROLLUP(col1, col2)
-		if p.currentToken.Type == "WITH" {
-			nextTok := p.peekToken()
-			if nextTok.Type == "ROLLUP" {
-				p.advance() // Consume WITH
-				p.advance() // Consume ROLLUP
-				// Wrap all existing expressions in a RollupExpression
-				groupByExprs = []ast.Expression{
-					&ast.RollupExpression{Expressions: groupByExprs},
-				}
-			} else if nextTok.Type == "CUBE" {
-				p.advance() // Consume WITH
-				p.advance() // Consume CUBE
-				// Wrap all existing expressions in a CubeExpression
-				groupByExprs = []ast.Expression{
-					&ast.CubeExpression{Expressions: groupByExprs},
-				}
-			}
-			// Note: WITH not followed by ROLLUP/CUBE will be handled elsewhere (e.g., CTE)
-		}
-
-		selectStmt.GroupBy = groupByExprs
-	}
-
-	// Parse HAVING clause if present (must come after GROUP BY)
-	if p.currentToken.Type == "HAVING" {
-		p.advance() // Consume HAVING
-
-		// Parse HAVING condition
-		havingClause, err := p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-		selectStmt.Having = havingClause
-	}
-
-	// Parse ORDER BY clause if present
-	if p.currentToken.Type == "ORDER" {
-		p.advance() // Consume ORDER
-
-		if p.currentToken.Type != "BY" {
-			return nil, p.expectedError("BY")
-		}
-		p.advance() // Consume BY
-
-		// Parse order expressions (comma-separated list)
-		for {
-			expr, err := p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-
-			// Create OrderByExpression with defaults
-			orderByExpr := ast.OrderByExpression{
-				Expression: expr,
-				Ascending:  true, // Default to ASC
-				NullsFirst: nil,  // Default behavior (database-specific)
-			}
-
-			// Check for ASC/DESC after the expression
-			if p.currentToken.Type == "ASC" {
-				orderByExpr.Ascending = true
-				p.advance() // Consume ASC
-			} else if p.currentToken.Type == "DESC" {
-				orderByExpr.Ascending = false
-				p.advance() // Consume DESC
-			}
-
-			// Check for NULLS FIRST/LAST
-			nullsFirst, err := p.parseNullsClause()
-			if err != nil {
-				return nil, err
-			}
-			orderByExpr.NullsFirst = nullsFirst
-
-			selectStmt.OrderBy = append(selectStmt.OrderBy, orderByExpr)
-
-			// Check for comma (more expressions)
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-			} else {
-				break
-			}
-		}
-	}
-
-	// Parse LIMIT clause if present
-	if p.currentToken.Type == "LIMIT" {
-		p.advance() // Consume LIMIT
-
-		// Parse LIMIT value
-		if p.currentToken.Type != "INT" {
-			return nil, p.expectedError("integer for LIMIT")
-		}
-
-		// Convert string to int
-		limitVal := 0
-		_, _ = fmt.Sscanf(p.currentToken.Literal, "%d", &limitVal)
-
-		// Add LIMIT to SELECT statement
-		selectStmt.Limit = &limitVal
-		p.advance()
-	}
-
-	// Parse OFFSET clause if present
-	if p.currentToken.Type == "OFFSET" {
-		p.advance() // Consume OFFSET
-
-		// Parse OFFSET value
-		if p.currentToken.Type != "INT" {
-			return nil, p.expectedError("integer for OFFSET")
-		}
-
-		// Convert string to int
-		offsetVal := 0
-		_, _ = fmt.Sscanf(p.currentToken.Literal, "%d", &offsetVal)
-
-		// Add OFFSET to SELECT statement
-		selectStmt.Offset = &offsetVal
-		p.advance()
-	}
-
-	return selectStmt, nil
-}
-
-// parseSelectWithSetOperations parses SELECT statements that may have set operations.
-// It supports UNION, UNION ALL, EXCEPT, and INTERSECT operations with proper left-associative parsing.
-//
-// Examples:
-//
-//	SELECT name FROM users UNION SELECT name FROM customers
-//	SELECT id FROM orders UNION ALL SELECT id FROM invoices
-//	SELECT product FROM inventory EXCEPT SELECT product FROM discontinued
-//	SELECT a FROM t1 UNION SELECT b FROM t2 INTERSECT SELECT c FROM t3
-func (p *Parser) parseSelectWithSetOperations() (ast.Statement, error) {
-	// Parse the first SELECT statement
-	leftStmt, err := p.parseSelectStatement()
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for set operations (UNION, EXCEPT, INTERSECT)
-	for p.currentToken.Type == "UNION" || p.currentToken.Type == "EXCEPT" || p.currentToken.Type == "INTERSECT" {
-		// Parse the set operation type
-		operationType := p.currentToken.Type
-		p.advance()
-
-		// Check for ALL keyword
-		all := false
-		if p.currentToken.Type == "ALL" {
-			all = true
-			p.advance()
-		}
-
-		// Parse the right-hand SELECT statement
-		if p.currentToken.Type != "SELECT" {
-			return nil, p.expectedError("SELECT after set operation")
-		}
-		p.advance() // Consume SELECT
-
-		rightStmt, err := p.parseSelectStatement()
-		if err != nil {
-			return nil, fmt.Errorf("error parsing right SELECT in set operation: %v", err)
-		}
-
-		// Create the set operation with left as the accumulated result
-		setOp := &ast.SetOperation{
-			Left:     leftStmt,
-			Operator: string(operationType),
-			All:      all,
-			Right:    rightStmt,
-		}
-
-		leftStmt = setOp // The result becomes the left side for any subsequent operations
-	}
-
-	return leftStmt, nil
-}
-
-// parseInsertStatement parses an INSERT statement
-func (p *Parser) parseInsertStatement() (ast.Statement, error) {
-	// We've already consumed the INSERT token in matchToken
-
-	// Parse INTO
-	if p.currentToken.Type != "INTO" {
-		return nil, p.expectedError("INTO")
-	}
-	p.advance() // Consume INTO
-
-	// Parse table name
-	if p.currentToken.Type != "IDENT" {
-		return nil, p.expectedError("table name")
-	}
-	tableName := p.currentToken.Literal
-	p.advance()
-
-	// Parse column list if present
-	columns := make([]ast.Expression, 0)
-	if p.currentToken.Type == "(" {
-		p.advance() // Consume (
-
-		for {
-			// Parse column name
-			if p.currentToken.Type != "IDENT" {
-				return nil, p.expectedError("column name")
-			}
-			columns = append(columns, &ast.Identifier{Name: p.currentToken.Literal})
-			p.advance()
-
-			// Check if there are more columns
-			if p.currentToken.Type != "," {
-				break
-			}
-			p.advance() // Consume comma
-		}
-
-		if p.currentToken.Type != ")" {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-	}
-
-	// Parse VALUES
-	if p.currentToken.Type != "VALUES" {
-		return nil, p.expectedError("VALUES")
-	}
-	p.advance() // Consume VALUES
-
-	// Parse value list
-	values := make([]ast.Expression, 0)
-	if p.currentToken.Type == "(" {
-		p.advance() // Consume (
-
-		for {
-			// Parse value
-			var expr ast.Expression
-			switch p.currentToken.Type {
-			case "STRING":
-				expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "string"}
-				p.advance()
-			case "INT":
-				expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "int"}
-				p.advance()
-			case "FLOAT":
-				expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "float"}
-				p.advance()
-			case token.TRUE, token.FALSE:
-				expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "bool"}
-				p.advance()
-			default:
-				return nil, fmt.Errorf("unexpected token for value: %s", p.currentToken.Type)
-			}
-			values = append(values, expr)
-
-			// Check if there are more values
-			if p.currentToken.Type != "," {
-				break
-			}
-			p.advance() // Consume comma
-		}
-
-		if p.currentToken.Type != ")" {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-	}
-
-	// Create INSERT statement
-	return &ast.InsertStatement{
-		TableName: tableName,
-		Columns:   columns,
-		Values:    values,
-	}, nil
-}
-
-// parseUpdateStatement parses an UPDATE statement
-func (p *Parser) parseUpdateStatement() (ast.Statement, error) {
-	// We've already consumed the UPDATE token in matchToken
-
-	// Parse table name
-	if p.currentToken.Type != "IDENT" {
-		return nil, p.expectedError("table name")
-	}
-	tableName := p.currentToken.Literal
-	p.advance()
-
-	// Parse SET
-	if p.currentToken.Type != "SET" {
-		return nil, p.expectedError("SET")
-	}
-	p.advance() // Consume SET
-
-	// Parse assignments
-	updates := make([]ast.UpdateExpression, 0)
-	for {
-		// Parse column name
-		if p.currentToken.Type != "IDENT" {
-			return nil, p.expectedError("column name")
-		}
-		columnName := p.currentToken.Literal
-		p.advance()
-
-		if p.currentToken.Type != "=" {
-			return nil, p.expectedError("=")
-		}
-		p.advance() // Consume =
-
-		// Parse value expression
-		var expr ast.Expression
-		switch p.currentToken.Type {
-		case "STRING":
-			expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "string"}
-			p.advance()
-		case "INT":
-			expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "int"}
-			p.advance()
-		case "FLOAT":
-			expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "float"}
-			p.advance()
-		case "TRUE", "FALSE":
-			expr = &ast.LiteralValue{Value: p.currentToken.Literal, Type: "bool"}
-			p.advance()
-		default:
-			var err error
-			expr, err = p.parseExpression()
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		// Create update expression
-		columnExpr := &ast.Identifier{Name: columnName}
-		updateExpr := ast.UpdateExpression{
-			Column: columnExpr,
-			Value:  expr,
-		}
-		updates = append(updates, updateExpr)
-
-		// Check if there are more assignments
-		if p.currentToken.Type != "," {
-			break
-		}
-		p.advance() // Consume comma
-	}
-
-	// Parse WHERE clause if present
-	var whereClause ast.Expression
-	if p.currentToken.Type == "WHERE" {
-		p.advance() // Consume WHERE
-		var err error
-		whereClause, err = p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Create UPDATE statement
-	return &ast.UpdateStatement{
-		TableName: tableName,
-		Updates:   updates,
-		Where:     whereClause,
-	}, nil
-}
-
-// parseDeleteStatement parses a DELETE statement
-func (p *Parser) parseDeleteStatement() (ast.Statement, error) {
-	// We've already consumed the DELETE token in matchToken
-
-	// Parse FROM
-	if p.currentToken.Type != "FROM" {
-		return nil, p.expectedError("FROM")
-	}
-	p.advance() // Consume FROM
-
-	// Parse table name
-	if p.currentToken.Type != "IDENT" {
-		return nil, p.expectedError("table name")
-	}
-	tableName := p.currentToken.Literal
-	p.advance()
-
-	// Parse WHERE clause if present
-	var whereClause ast.Expression
-	if p.currentToken.Type == "WHERE" {
-		p.advance() // Consume WHERE
-		var err error
-		whereClause, err = p.parseExpression()
-		if err != nil {
-			return nil, err
-		}
-	}
-
-	// Create DELETE statement
-	return &ast.DeleteStatement{
-		TableName: tableName,
-		Where:     whereClause,
-	}, nil
-}
-
-// parseMergeStatement parses a MERGE statement (SQL:2003 F312)
-// Syntax: MERGE INTO target [AS alias] USING source [AS alias] ON condition
-//
-//	WHEN MATCHED [AND condition] THEN UPDATE/DELETE
-//	WHEN NOT MATCHED [AND condition] THEN INSERT
-//	WHEN NOT MATCHED BY SOURCE [AND condition] THEN UPDATE/DELETE
-func (p *Parser) parseMergeStatement() (ast.Statement, error) {
-	stmt := &ast.MergeStatement{}
-
-	// Parse INTO (optional)
-	if p.currentToken.Type == "INTO" {
-		p.advance() // Consume INTO
-	}
-
-	// Parse target table
-	tableRef, err := p.parseTableReference()
-	if err != nil {
-		return nil, fmt.Errorf("error parsing MERGE target table: %w", err)
-	}
-	stmt.TargetTable = *tableRef
-
-	// Parse optional target alias (AS alias or just alias)
-	if p.currentToken.Type == "AS" {
-		p.advance() // Consume AS
-		if p.currentToken.Type != "IDENT" && !p.isNonReservedKeyword() {
-			return nil, p.expectedError("target alias after AS")
-		}
-		stmt.TargetAlias = p.currentToken.Literal
-		p.advance()
-	} else if p.canBeAlias() && p.currentToken.Type != "USING" && p.currentToken.Literal != "USING" {
-		stmt.TargetAlias = p.currentToken.Literal
-		p.advance()
-	}
-
-	// Parse USING
-	if p.currentToken.Type != "USING" && p.currentToken.Literal != "USING" {
-		return nil, p.expectedError("USING")
-	}
-	p.advance() // Consume USING
-
-	// Parse source table (could be a table or subquery)
-	sourceRef, err := p.parseTableReference()
-	if err != nil {
-		return nil, fmt.Errorf("error parsing MERGE source: %w", err)
-	}
-	stmt.SourceTable = *sourceRef
-
-	// Parse optional source alias
-	if p.currentToken.Type == "AS" {
-		p.advance() // Consume AS
-		if p.currentToken.Type != "IDENT" && !p.isNonReservedKeyword() {
-			return nil, p.expectedError("source alias after AS")
-		}
-		stmt.SourceAlias = p.currentToken.Literal
-		p.advance()
-	} else if p.canBeAlias() && p.currentToken.Type != "ON" && p.currentToken.Literal != "ON" {
-		stmt.SourceAlias = p.currentToken.Literal
-		p.advance()
-	}
-
-	// Parse ON condition
-	if p.currentToken.Type != "ON" {
-		return nil, p.expectedError("ON")
-	}
-	p.advance() // Consume ON
-
-	onCondition, err := p.parseExpression()
-	if err != nil {
-		return nil, fmt.Errorf("error parsing MERGE ON condition: %w", err)
-	}
-	stmt.OnCondition = onCondition
-
-	// Parse WHEN clauses
-	for p.currentToken.Type == "WHEN" {
-		whenClause, err := p.parseMergeWhenClause()
-		if err != nil {
-			return nil, err
-		}
-		stmt.WhenClauses = append(stmt.WhenClauses, whenClause)
-	}
-
-	if len(stmt.WhenClauses) == 0 {
-		return nil, fmt.Errorf("MERGE statement requires at least one WHEN clause")
-	}
-
-	return stmt, nil
-}
-
-// parseMergeWhenClause parses a WHEN clause in a MERGE statement
-func (p *Parser) parseMergeWhenClause() (*ast.MergeWhenClause, error) {
-	clause := &ast.MergeWhenClause{}
-
-	p.advance() // Consume WHEN
-
-	// Determine clause type: MATCHED, NOT MATCHED, NOT MATCHED BY SOURCE
-	if p.currentToken.Type == "MATCHED" || p.currentToken.Literal == "MATCHED" {
-		clause.Type = "MATCHED"
-		p.advance() // Consume MATCHED
-	} else if p.currentToken.Type == "NOT" {
-		p.advance() // Consume NOT
-		if p.currentToken.Type != "MATCHED" && p.currentToken.Literal != "MATCHED" {
-			return nil, p.expectedError("MATCHED after NOT")
-		}
-		p.advance() // Consume MATCHED
-
-		// Check for BY SOURCE
-		if p.currentToken.Type == "BY" {
-			p.advance() // Consume BY
-			if p.currentToken.Type != "SOURCE" && p.currentToken.Literal != "SOURCE" {
-				return nil, p.expectedError("SOURCE after BY")
-			}
-			p.advance() // Consume SOURCE
-			clause.Type = "NOT_MATCHED_BY_SOURCE"
-		} else {
-			clause.Type = "NOT_MATCHED"
-		}
-	} else {
-		return nil, p.expectedError("MATCHED or NOT MATCHED")
-	}
-
-	// Parse optional AND condition
-	if p.currentToken.Type == "AND" {
-		p.advance() // Consume AND
-		condition, err := p.parseExpression()
-		if err != nil {
-			return nil, fmt.Errorf("error parsing WHEN condition: %w", err)
-		}
-		clause.Condition = condition
-	}
-
-	// Parse THEN
-	if p.currentToken.Type != "THEN" {
-		return nil, p.expectedError("THEN")
-	}
-	p.advance() // Consume THEN
-
-	// Parse action (UPDATE, INSERT, DELETE)
-	action, err := p.parseMergeAction(clause.Type)
-	if err != nil {
-		return nil, err
-	}
-	clause.Action = action
-
-	return clause, nil
-}
-
-// parseMergeAction parses the action in a WHEN clause
-func (p *Parser) parseMergeAction(clauseType string) (*ast.MergeAction, error) {
-	action := &ast.MergeAction{}
-
-	switch p.currentToken.Type {
-	case "UPDATE":
-		action.ActionType = "UPDATE"
-		p.advance() // Consume UPDATE
-
-		// Parse SET
-		if p.currentToken.Type != "SET" {
-			return nil, p.expectedError("SET after UPDATE")
-		}
-		p.advance() // Consume SET
-
-		// Parse SET clauses
-		for {
-			if p.currentToken.Type != "IDENT" && !p.canBeAlias() {
-				return nil, p.expectedError("column name")
-			}
-			// Handle qualified column names (e.g., t.name)
-			columnName := p.currentToken.Literal
-			p.advance()
-
-			// Check for qualified name (table.column)
-			if p.currentToken.Type == "." {
-				p.advance() // Consume .
-				if p.currentToken.Type != "IDENT" && !p.canBeAlias() {
-					return nil, p.expectedError("column name after .")
-				}
-				columnName = columnName + "." + p.currentToken.Literal
-				p.advance()
-			}
-
-			setClause := ast.SetClause{Column: columnName}
-
-			if p.currentToken.Type != "=" {
-				return nil, p.expectedError("=")
-			}
-			p.advance() // Consume =
-
-			value, err := p.parseExpression()
-			if err != nil {
-				return nil, fmt.Errorf("error parsing SET value: %w", err)
-			}
-			setClause.Value = value
-			action.SetClauses = append(action.SetClauses, setClause)
-
-			if p.currentToken.Type != "," {
-				break
-			}
-			p.advance() // Consume comma
-		}
-
-	case "INSERT":
-		if clauseType == "MATCHED" || clauseType == "NOT_MATCHED_BY_SOURCE" {
-			return nil, fmt.Errorf("INSERT not allowed in WHEN %s clause", clauseType)
-		}
-		action.ActionType = "INSERT"
-		p.advance() // Consume INSERT
-
-		// Parse optional column list
-		if p.currentToken.Type == "(" {
-			p.advance() // Consume (
-			for {
-				if p.currentToken.Type != "IDENT" {
-					return nil, p.expectedError("column name")
-				}
-				action.Columns = append(action.Columns, p.currentToken.Literal)
-				p.advance()
-
-				if p.currentToken.Type != "," {
-					break
-				}
-				p.advance() // Consume comma
-			}
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-		}
-
-		// Parse VALUES or DEFAULT VALUES
-		if p.currentToken.Type == "DEFAULT" {
-			p.advance() // Consume DEFAULT
-			if p.currentToken.Type != "VALUES" {
-				return nil, p.expectedError("VALUES after DEFAULT")
-			}
-			p.advance() // Consume VALUES
-			action.DefaultValues = true
-		} else if p.currentToken.Type == "VALUES" {
-			p.advance() // Consume VALUES
-			if p.currentToken.Type != "(" {
-				return nil, p.expectedError("(")
-			}
-			p.advance() // Consume (
-
-			for {
-				value, err := p.parseExpression()
-				if err != nil {
-					return nil, fmt.Errorf("error parsing INSERT value: %w", err)
-				}
-				action.Values = append(action.Values, value)
-
-				if p.currentToken.Type != "," {
-					break
-				}
-				p.advance() // Consume comma
-			}
-
-			if p.currentToken.Type != ")" {
-				return nil, p.expectedError(")")
-			}
-			p.advance() // Consume )
-		} else {
-			return nil, p.expectedError("VALUES or DEFAULT VALUES")
-		}
-
-	case "DELETE":
-		if clauseType == "NOT_MATCHED" {
-			return nil, fmt.Errorf("DELETE not allowed in WHEN NOT MATCHED clause")
-		}
-		action.ActionType = "DELETE"
-		p.advance() // Consume DELETE
-
-	default:
-		return nil, p.expectedError("UPDATE, INSERT, or DELETE")
-	}
-
-	return action, nil
-}
-
-// parseTableReference parses a simple table reference (table name)
-// Returns a TableReference with the Name field populated
 // Accepts IDENT or non-reserved keywords that can be used as table names
 func (p *Parser) parseTableReference() (*ast.TableReference, error) {
 	// Accept IDENT or keywords that can be used as table names
@@ -2358,166 +353,3 @@ func (p *Parser) isJoinKeyword() bool {
 //	WITH RECURSIVE emp_tree AS (SELECT emp_id FROM employees) SELECT * FROM emp_tree
 //	WITH first AS (SELECT * FROM t1), second AS (SELECT * FROM first) SELECT * FROM second
 //	WITH summary(region, total) AS (SELECT region, SUM(amount) FROM sales GROUP BY region) SELECT * FROM summary
-func (p *Parser) parseWithStatement() (ast.Statement, error) {
-	// Consume WITH
-	p.advance()
-
-	// Check for RECURSIVE keyword
-	recursive := false
-	if p.currentToken.Type == "RECURSIVE" {
-		recursive = true
-		p.advance()
-	}
-
-	// Parse Common Table Expressions
-	ctes := []*ast.CommonTableExpr{}
-
-	for {
-		cte, err := p.parseCommonTableExpr()
-		if err != nil {
-			return nil, fmt.Errorf("error parsing CTE: %v", err)
-		}
-		ctes = append(ctes, cte)
-
-		// Check for more CTEs (comma-separated)
-		if p.currentToken.Type == "," {
-			p.advance() // Consume comma
-			continue
-		}
-		break
-	}
-
-	// Create WITH clause
-	withClause := &ast.WithClause{
-		Recursive: recursive,
-		CTEs:      ctes,
-	}
-
-	// Parse the main statement that follows the WITH clause
-	mainStmt, err := p.parseMainStatementAfterWith()
-	if err != nil {
-		return nil, fmt.Errorf("error parsing statement after WITH: %v", err)
-	}
-
-	// Attach WITH clause to the main statement
-	switch stmt := mainStmt.(type) {
-	case *ast.SelectStatement:
-		stmt.With = withClause
-		return stmt, nil
-	case *ast.SetOperation:
-		// For set operations, attach WITH to the left statement if it's a SELECT
-		if leftSelect, ok := stmt.Left.(*ast.SelectStatement); ok {
-			leftSelect.With = withClause
-		}
-		return stmt, nil
-	case *ast.InsertStatement:
-		stmt.With = withClause
-		return stmt, nil
-	case *ast.UpdateStatement:
-		stmt.With = withClause
-		return stmt, nil
-	case *ast.DeleteStatement:
-		stmt.With = withClause
-		return stmt, nil
-	default:
-		return nil, fmt.Errorf("WITH clause not supported with statement type: %T", stmt)
-	}
-}
-
-// parseCommonTableExpr parses a single Common Table Expression.
-// It handles CTE name, optional column list, AS keyword, and the CTE query in parentheses.
-//
-// Syntax: cte_name [(column_list)] AS (query)
-func (p *Parser) parseCommonTableExpr() (*ast.CommonTableExpr, error) {
-	// Check recursion depth to prevent stack overflow in recursive CTEs
-	// This is critical since CTEs can call parseStatement which leads back to more CTEs
-	p.depth++
-	defer func() { p.depth-- }()
-
-	if p.depth > MaxRecursionDepth {
-		return nil, fmt.Errorf("maximum recursion depth exceeded (%d) - CTE too deeply nested", MaxRecursionDepth)
-	}
-
-	// Parse CTE name
-	if p.currentToken.Type != "IDENT" {
-		return nil, p.expectedError("CTE name")
-	}
-	name := p.currentToken.Literal
-	p.advance()
-
-	// Parse optional column list
-	var columns []string
-	if p.currentToken.Type == "(" {
-		p.advance() // Consume (
-
-		for {
-			if p.currentToken.Type != "IDENT" {
-				return nil, p.expectedError("column name")
-			}
-			columns = append(columns, p.currentToken.Literal)
-			p.advance()
-
-			if p.currentToken.Type == "," {
-				p.advance() // Consume comma
-				continue
-			}
-			break
-		}
-
-		if p.currentToken.Type != ")" {
-			return nil, p.expectedError(")")
-		}
-		p.advance() // Consume )
-	}
-
-	// Parse AS keyword
-	if p.currentToken.Type != "AS" {
-		return nil, p.expectedError("AS")
-	}
-	p.advance()
-
-	// Parse the CTE query (must be in parentheses)
-	if p.currentToken.Type != "(" {
-		return nil, p.expectedError("( before CTE query")
-	}
-	p.advance() // Consume (
-
-	// Parse the inner statement
-	stmt, err := p.parseStatement()
-	if err != nil {
-		return nil, fmt.Errorf("error parsing CTE statement: %v", err)
-	}
-
-	if p.currentToken.Type != ")" {
-		return nil, p.expectedError(") after CTE query")
-	}
-	p.advance() // Consume )
-
-	return &ast.CommonTableExpr{
-		Name:      name,
-		Columns:   columns,
-		Statement: stmt,
-	}, nil
-}
-
-// parseMainStatementAfterWith parses the main statement after WITH clause.
-// It supports SELECT, INSERT, UPDATE, and DELETE statements, routing them to the appropriate
-// parsers while preserving set operation support for SELECT statements.
-func (p *Parser) parseMainStatementAfterWith() (ast.Statement, error) {
-	switch p.currentToken.Type {
-	case "SELECT":
-		p.advance() // Consume SELECT
-		return p.parseSelectWithSetOperations()
-	case "INSERT":
-		p.advance() // Consume INSERT
-		return p.parseInsertStatement()
-	case "UPDATE":
-		p.advance() // Consume UPDATE
-		return p.parseUpdateStatement()
-	case "DELETE":
-		p.advance() // Consume DELETE
-		return p.parseDeleteStatement()
-	default:
-		return nil, p.expectedError("SELECT, INSERT, UPDATE, or DELETE after WITH")
-	}
-}
